remove json-server
