--- boilerplates/redux/json-server.json
@@ -1,106 +1,0 @@
-{
-  "users": [
-    {
-      "id": 1,
-      "name": "胡杰",
-      "age": 54,
-      "address": "华东"
-    },
-    {
-      "id": 2,
-      "name": "沈娟",
-      "age": 14,
-      "address": "华南"
-    },
-    {
-      "id": 3,
-      "name": "薛娟",
-      "age": 30,
-      "address": "华北"
-    },
-    {
-      "id": 4,
-      "name": "蔡平",
-      "age": 42,
-      "address": "华东"
-    },
-    {
-      "id": 5,
-      "name": "夏杰",
-      "age": 21,
-      "address": "西北"
-    },
-    {
-      "id": 6,
-      "name": "雷杰",
-      "age": 27,
-      "address": "华东"
-    },
-    {
-      "id": 7,
-      "name": "万强",
-      "age": 70,
-      "address": "华中"
-    },
-    {
-      "id": 8,
-      "name": "丁刚",
-      "age": 13,
-      "address": "华北"
-    },
-    {
-      "id": 9,
-      "name": "李艳",
-      "age": 96,
-      "address": "华中"
-    },
-    {
-      "id": 10,
-      "name": "史平",
-      "age": 61,
-      "address": "华东"
-    },
-    {
-      "id": 11,
-      "name": "孙军",
-      "age": 14,
-      "address": "西南"
-    },
-    {
-      "id": 12,
-      "name": "孙桂英",
-      "age": 90,
-      "address": "华中"
-    },
-    {
-      "id": 13,
-      "name": "谭霞",
-      "age": 83,
-      "address": "华南"
-    },
-    {
-      "id": 14,
-      "name": "傅娜",
-      "age": 64,
-      "address": "华中"
-    },
-    {
-      "id": 15,
-      "name": "江霞",
-      "age": 58,
-      "address": "西南"
-    },
-    {
-      "id": 16,
-      "name": "陈秀英",
-      "age": 24,
-      "address": "华东"
-    },
-    {
-      "id": 17,
-      "name": "汪杰",
-      "age": 23,
-      "address": "西北"
-    }
-  ]
-}
