+++ boilerplates/redux/package.json
@@ -26,7 +26,6 @@
     "atool-test-mocha": "^0.1.4",
     "babel-eslint": "^6.0.2",
     "dora": "0.3.x",
-    "dora-plugin-json-server": "^0.1.1",
     "dora-plugin-livereload": "^0.4.0",
     "dora-plugin-proxy": "^0.7.0",
     "dora-plugin-webpack": "0.6.x",
